Implemented TimeJudge view binding

+++ app/src/main/java/com/imuaythai/mtjudges/provider/dto/FightStatusDto.kt
@@ -3,5 +3,5 @@ package com.imuaythai.mtjudges.provider.dto
 data class FightStatusDto(
     val roundNum: Int,
     val state: FightState,
-    val time: Int
+    val time: Long
 )
